Committing the refactoring changes for the project
