+++ core/src/main/java/ma/glasnost/orika/impl/generator/AggregateSpecification.java
@@ -32,7 +32,7 @@ import ma.glasnost.orika.metadata.FieldMap;
  */
 public interface AggregateSpecification extends BaseSpecification {
     
-    void setMapperFactory(MapperFactory mapperFactory);
+
     
     /**
      * @param fieldMappings
